third_party/blink/web_tests/platform/generic/http/tests/misc/script-sync-slow-scripts-onerror-expected.txt

Slow scripts that fail to load should not dispatch multiple error events

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS firstOnErrorHandlerCount is 1
PASS secondOnErrorHandlerCount is 1
PASS successfullyParsed is true

TEST COMPLETE
